use same copy behavior as for journal abbrevs
